Update the packing script

+++ pack-and-install.ps1
@@ -1,7 +1,17 @@
+if (-not (Test-Path .env)) {
+    Write-Error "Please copy .env.in and set the values inside"
+    exit 1
+}
+Get-Content .env | Foreach-Object {
+    $name, $value = $_.split('=')
+    Set-Variable $name $value -Scope Local
+}
+
 $VERSION = (Get-Content -Raw info.json | ConvertFrom-Json).version
+$MODNAME = (Get-Content -Raw info.json | ConvertFrom-Json).name
+$PACKAGE_NAME = "${MODNAME}_${VERSION}"
 
-$TMPDIR = "builds/cargo-icbm_$VERSION"
-
+$TMPDIR = "builds/$PACKAGE_NAME"
 New-Item -ItemType "Directory" -Force $TMPDIR
 Remove-Item -Recurse $TMPDIR/*
 
@@ -9,7 +19,8 @@ Copy-Item info.json $TMPDIR
 Copy-Item control.lua $TMPDIR
 Copy-Item thumbnail.png $TMPDIR
 
-Compress-Archive -Path $TMPDIR -DestinationPath "$TMPDIR.zip" -Force
-# TODO: replace with 7zip to avoid backslashes in paths, which Factorio.com doesn't accept
+cd builds
+&$7ZIP_BINARY a -r "$PACKAGE_NAME.zip" $PACKAGE_NAME
+cd ..
 
-Copy-Item "$TMPDIR.zip" C:\Users\alexa\AppData\Roaming\Factorio\mods\
+Copy-Item "builds/$PACKAGE_NAME.zip" $LOCAL_MODFOLDER
